+++ src/Shortcut/Wechat/MiniShortcut.php
@@ -2,22 +2,24 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\Shortcut;
+namespace Yansongda\Pay\Shortcut\Wechat;
 
 use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Wechat\LaunchPlugin;
-use Yansongda\Pay\Plugin\Wechat\Pay\H5\PrepayPlugin;
+use Yansongda\Pay\Plugin\Wechat\Pay\Mini\InvokePrepayPlugin;
+use Yansongda\Pay\Plugin\Wechat\Pay\Mini\PrepayPlugin;
 use Yansongda\Pay\Plugin\Wechat\PreparePlugin;
 use Yansongda\Pay\Plugin\Wechat\RadarSignPlugin;
 
-class WapShortcut implements ShortcutInterface
+class MiniShortcut implements ShortcutInterface
 {
     public function getPlugins(array $params): array
     {
         return [
             PreparePlugin::class,
             PrepayPlugin::class,
+            InvokePrepayPlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,