+++ src/Shortcut/Wechat/PosShortcut.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\Shortcut;
+namespace Yansongda\Pay\Shortcut\Wechat;
 
 use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Plugin\Wechat\Pay\Pos\PayPlugin;