+++ src/Shortcut/Wechat/WapShortcut.php
@@ -2,24 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Plugin\Wechat\Shortcut;
+namespace Yansongda\Pay\Shortcut\Wechat;
 
 use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Wechat\LaunchPlugin;
-use Yansongda\Pay\Plugin\Wechat\Pay\App\InvokePrepayPlugin;
-use Yansongda\Pay\Plugin\Wechat\Pay\App\PrepayPlugin;
+use Yansongda\Pay\Plugin\Wechat\Pay\H5\PrepayPlugin;
 use Yansongda\Pay\Plugin\Wechat\PreparePlugin;
 use Yansongda\Pay\Plugin\Wechat\RadarSignPlugin;
 
-class AppShortcut implements ShortcutInterface
+class WapShortcut implements ShortcutInterface
 {
     public function getPlugins(array $params): array
     {
         return [
             PreparePlugin::class,
             PrepayPlugin::class,
-            InvokePrepayPlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,