+++ tests/Shortcut/Wechat/CloseShortcutTest.php
@@ -2,57 +2,46 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Tests\Plugin\Wechat\Shortcut;
+namespace Yansongda\Pay\Tests\Shortcut\Wechat;
 
 use Yansongda\Pay\Exception\Exception;
 use Yansongda\Pay\Exception\InvalidParamsException;
+use Yansongda\Pay\Plugin\Alipay\Fund\TransCommonQueryPlugin;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Wechat\LaunchPlugin;
-use Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryRefundPlugin;
-use Yansongda\Pay\Plugin\Wechat\Pay\Common\QueryPlugin;
+use Yansongda\Pay\Plugin\Wechat\Pay\Common\ClosePlugin;
 use Yansongda\Pay\Plugin\Wechat\PreparePlugin;
 use Yansongda\Pay\Plugin\Wechat\RadarSignPlugin;
-use Yansongda\Pay\Plugin\Wechat\Shortcut\QueryShortcut;
+use Yansongda\Pay\Shortcut\Wechat\CloseShortcut;
 use Yansongda\Pay\Tests\TestCase;
 
-class QueryShortcutTest extends TestCase
+class CloseShortcutTest extends TestCase
 {
-    protected QueryShortcut $plugin;
+    protected CloseShortcut $plugin;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->plugin = new QueryShortcut();
+        $this->plugin = new CloseShortcut();
     }
 
     public function testDefault()
     {
         self::assertEquals([
             PreparePlugin::class,
-            QueryPlugin::class,
+            ClosePlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
         ], $this->plugin->getPlugins([]));
     }
 
-    public function testRefund()
-    {
-        self::assertEquals([
-            PreparePlugin::class,
-            QueryRefundPlugin::class,
-            RadarSignPlugin::class,
-            LaunchPlugin::class,
-            ParserPlugin::class,
-        ], $this->plugin->getPlugins(['_action' => 'refund']));
-    }
-
     public function testCombine()
     {
         self::assertEquals([
             PreparePlugin::class,
-            \Yansongda\Pay\Plugin\Wechat\Pay\Combine\QueryPlugin::class,
+            \Yansongda\Pay\Plugin\Wechat\Pay\Combine\ClosePlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
@@ -63,11 +52,19 @@ class QueryShortcutTest extends TestCase
     {
         self::assertEquals([
             PreparePlugin::class,
-            \Yansongda\Pay\Plugin\Wechat\Pay\Combine\QueryPlugin::class,
+            \Yansongda\Pay\Plugin\Wechat\Pay\Combine\ClosePlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
         ], $this->plugin->getPlugins(['combine_out_trade_no' => '123abc']));
+
+        self::assertEquals([
+            PreparePlugin::class,
+            \Yansongda\Pay\Plugin\Wechat\Pay\Combine\ClosePlugin::class,
+            RadarSignPlugin::class,
+            LaunchPlugin::class,
+            ParserPlugin::class,
+        ], $this->plugin->getPlugins(['sub_orders' => '123abc']));
     }
 
     public function testFoo()