+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - change: cancel 的 API 参数只支持 array，不再支持 string(#900)
 - change: close 的 API 参数只支持 array，不再支持 string(#901)
 - change(internal): 错误代码 const 命名规则统一(#902, #903)
+- change(internal): 微信 shortcut 从 plugin 文件夹独立出来(#904)
 
 ## v3.5.3